// src/utils/types.ts
export type Data = {
    menu: { id: string; name: string }[];
    logo: string;
    backgroundImage: string;
    header: {
        mainTitle: string;
        title: string;
        headersButton: string;
    };
    benefits: {
        id: string;
        firstLine: string;
        secondLine: string;
        thirdLine: string;
    }[];
};